refactor: redistribute legacy datatypes
